+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## next
+### Enhancements
+- return `availableItems` property from `useFillItem()`. This contains the id-s of items that are
+actually available (visible) in the session.
 
 ## 2.0.1
 ### Bug fixes